Move trafficmanager/2018-02-01 to preview since this should be a preview package
