+++ services/preview/trafficmanager/mgmt/2018-02-01-preview/trafficmanager/enums.go
